Factor CREATE_IS_CALLABLE into its own file
