+++ src/DataStructures/DataBox/DataBox.hpp
@@ -28,6 +28,7 @@
 #include "Utilities/Requires.hpp"
 #include "Utilities/TMPL.hpp"
 #include "Utilities/TypeTraits.hpp"
+#include "Utilities/TypeTraits/CreateIsCallable.hpp"
 
 // IWYU pragma: no_forward_declare brigand::get_destination
 // IWYU pragma: no_forward_declare brigand::get_source
