+++ src/DataStructures/DataBox/DataBoxTag.hpp
@@ -21,6 +21,7 @@
 #include "Utilities/Requires.hpp"
 #include "Utilities/TMPL.hpp"
 #include "Utilities/TypeTraits.hpp"
+#include "Utilities/TypeTraits/CreateIsCallable.hpp"
 
 namespace Tags {
 /*!
