+++ src/Domain/CoordinateMaps/TimeDependentHelpers.hpp
@@ -12,6 +12,7 @@
 
 #include "Domain/FunctionsOfTime/FunctionOfTime.hpp"
 #include "Utilities/TypeTraits.hpp"
+#include "Utilities/TypeTraits/CreateIsCallable.hpp"
 
 namespace domain {
 /// Check if the calls to the coordinate map and its inverse map are
