+++ src/Parallel/AlgorithmMetafunctions.hpp
@@ -10,7 +10,7 @@
 #include "Parallel/PhaseDependentActionList.hpp"
 #include "Utilities/TMPL.hpp"
 #include "Utilities/TaggedTuple.hpp"
-#include "Utilities/TypeTraits.hpp"
+#include "Utilities/TypeTraits/CreateIsCallable.hpp"
 
 namespace Parallel {
 /// \cond
