+++ tests/Unit/Utilities/Test_ProtocolHelpers.cpp
@@ -8,6 +8,7 @@
 #include "Utilities/ProtocolHelpers.hpp"
 #include "Utilities/TypeTraits.hpp"
 #include "Utilities/TypeTraits/CreateHasTypeAlias.hpp"
+#include "Utilities/TypeTraits/CreateIsCallable.hpp"
 #include "tests/Unit/ProtocolTestHelpers.hpp"
 
 namespace {
