+++ tests/Unit/Utilities/TypeTraits/CMakeLists.txt
@@ -6,6 +6,7 @@ set(LIBRARY "Test_TypeTraits")
 set(LIBRARY_SOURCES
   Test_CreateHasStaticMemberVariable.cpp
   Test_CreateHasTypeAlias.cpp
+  Test_CreateIsCallable.cpp
   Test_FunctionInfo.cpp
   Test_GetFundamentalType.cpp
   Test_IsComplexOfFundamental.cpp
